BigQueryドライバーのクラス分離とドライバー登録修正
- AdminerLoginBigQuery、AdminerBigQueryCSS、BigQueryUtilsクラスを独立ファイルに分離
- OAuth2認証機能をAdminerLoginBigQueryクラスに移動
- CSS機能をAdminerBigQueryCSSクラスに移動
- ユーティリティ機能をBigQueryUtilsクラスに移動
- plugins/drivers/bigquery/ディレクトリに配置、namespace Adminer使用
- ドライバー登録の条件チェックを削除してAdminerドロップダウンに正常表示
- /test-webコマンドで動作確認済み

修正により、BigQueryドライバーがAdminerインターフェースで正常選択可能
