+++ devtools/issues/i06.md
@@ -31,8 +31,10 @@
 
 # 3. クラス分離
 * bigquery.php に含まれているクラスをファイルに分割します。
-* 1クラスつづ作業します。今回はBigQueryConfigクラスを分離します。
+* 1クラスつづ作業します。今回はAdminerLoginBigQueryクラスを分離します。
 * 分離されたファイルはplugins/drivers/bigquery/に配置してください
+* namespace Adminer; であることに注意してください
+* 分離されたら、
 
 
 # 4. SQL発行メソッドの共通化
@@ -40,3 +42,11 @@
 * SQLを受け取って実行するメソッドを適切なクラス内に作成し、グローバルfunctionではSQL文をそのメソッドに渡すだけにしてください。
 * adminer/drivers/mysql.inc.php を参考にしてください。
 * bigquery.php内のすべてのfunctionについて、同様の冗長コード排除を行ってください
+
+
+# 3. クラス分離ソノ２　
+* bigquery.php に含まれているクラスをファイルに分割します。
+* 1クラスつづ作業します。今回はAdminerLoginBigQueryクラスを分離します。
+* 分離されたファイルはplugins/drivers/bigquery/に配置してください
+* namespace Adminer; であることに注意してください
+* 分離されたら、/test-webコマンドでテストしてください
