+++ plugins/drivers/bigquery.php
@@ -6,573 +6,25 @@ use Google\Cloud\BigQuery\BigQueryClient;
 use Google\Cloud\Core\Exception\ServiceException;
 use Exception;
 use InvalidArgumentException;
+require_once __DIR__ . '/bigquery/AdminerLoginBigQuery.php';
+require_once __DIR__ . '/bigquery/AdminerBigQueryCSS.php';
+require_once __DIR__ . '/bigquery/BigQueryCacheManager.php';
+require_once __DIR__ . '/bigquery/BigQueryConnectionPool.php';
+require_once __DIR__ . '/bigquery/BigQueryConfig.php';
 
-if (function_exists('Adminer\\add_driver')) {
-	add_driver("bigquery", "Google BigQuery");
-}
 
-class AdminerLoginBigQuery extends Plugin
-{
-	protected $config;
-
-	function __construct($config = array())
-	{
-		$this->config = $config;
-		$this->initializeDriverSelection();
-	}
-	/**
-	 * OAuth2認証が有効かどうかをチェック
-	 */
-	private function isOAuth2Enabled()
-	{
-		$oauth2Enable = getenv('GOOGLE_OAUTH2_ENABLE');
-		// Add input validation and sanitization
-		return $oauth2Enable === 'true' && $this->validateOAuth2Configuration();
-	}
-
-	/**
-	 * Validate OAuth2 configuration for security
-	 */
-	private function validateOAuth2Configuration()
-	{
-		$clientId = getenv('GOOGLE_OAUTH2_CLIENT_ID');
-		$redirectUrl = getenv('GOOGLE_OAUTH2_REDIRECT_URL');
-
-		// Validate client ID format
-		if (!$clientId || !preg_match('/^[a-zA-Z0-9\-_.]+$/', $clientId)) {
-			error_log('OAuth2: Invalid client ID format');
-			return false;
-		}
-
-		// Validate redirect URL
-		if (!$redirectUrl || !filter_var($redirectUrl, FILTER_VALIDATE_URL)) {
-			error_log('OAuth2: Invalid redirect URL');
-			return false;
-		}
-
-		// Ensure HTTPS for production
-		if (parse_url($redirectUrl, PHP_URL_SCHEME) !== 'https' &&
-			!in_array($_SERVER['HTTP_HOST'] ?? '', ['localhost', '127.0.0.1'])) {
-			error_log('OAuth2: Redirect URL must use HTTPS in production');
-			return false;
-		}
-
-		return true;
-	}
-
-	/**
-	 * Validate OAuth2 state parameter for CSRF protection
-	 */
-	private function validateStateParameter($state)
-	{
-		try {
-			$decodedState = base64_decode($state, true);
-			if ($decodedState === false) {
-				return false;
-			}
-
-			$stateData = json_decode($decodedState, true);
-			if (!is_array($stateData)) {
-				return false;
-			}
-
-			// Validate redirect_to parameter
-			if (isset($stateData['redirect_to'])) {
-				$redirectTo = $stateData['redirect_to'];
-				// Only allow relative URLs or same-origin URLs
-				if (!preg_match('/^\/[^\/]/', $redirectTo) &&
-					parse_url($redirectTo, PHP_URL_HOST) !== $_SERVER['HTTP_HOST']) {
-					return false;
-				}
-			}
-
-			return true;
-		} catch (Exception $e) {
-			error_log('OAuth2: State validation error: ' . $e->getMessage());
-			return false;
-		}
-	}
-
-	/**
-	 * OAuth2設定を取得
-	 */
-	private function getOAuth2Config()
-	{
-		return [
-			'client_id' => getenv('GOOGLE_OAUTH2_CLIENT_ID'),
-			'redirect_url' => getenv('GOOGLE_OAUTH2_REDIRECT_URL')
-		];
-	}
-
-	private function initializeDriverSelection()
-	{
-		if (!isset($_POST["auth"])) {
-			return;
-		}
-
-		$_POST["auth"]["driver"] = 'bigquery';
-	}
-
-	function credentials()
-	{
-		$server = $this->getProjectId();
-
-		return array($server, 'bigquery-service-account', 'service-account-auth');
-	}
-
-	private function getProjectId()
-	{
-		return $_GET["server"] ??
-			$_POST["auth"]["server"] ??
-			$this->config['project_id'];
-	}
-
-	function login($login, $password)
-	{
-		return true;
-	}
-
-	function loginFormField($name, $heading, $value)
-	{
-		// OAuth2認証が有効な場合は異なる表示
-		if ($this->isOAuth2Enabled()) {
-			return $this->renderOAuth2Field($name, $heading, $value);
-		}
-
-		// 従来のCREDENTIAL認証
-		$fieldHandlers = array(
-			'driver' => function () use ($heading) {
-				return $this->renderDriverField($heading);
-			},
-			'server' => function () {
-				return $this->renderProjectIdField();
-			},
-			'username' => function () {
-				return $this->renderHiddenField('username');
-			},
-			'password' => function () {
-				return $this->renderHiddenField('password');
-			},
-			'db' => function () {
-				return $this->renderHiddenField('db');
-			}
-		);
-
-		return isset($fieldHandlers[$name]) ? $fieldHandlers[$name]() : '';
-	}
-
-	private function renderDriverField($heading)
-	{
-		return $heading . '<select name="auth[driver]" readonly><option value="bigquery" selected>Google BigQuery</option></select>' . "\n";
-	}
-
-	private function renderProjectIdField()
-	{
-		$default_value = htmlspecialchars($this->getProjectId());
-		return '<tr><th>Project ID</th><td><input name="auth[server]" value="' . $default_value . '" title="GCP Project ID" placeholder="your-project-id" autocapitalize="off" required></td></tr>' . "\n";
-	}
-
-	private function renderHiddenField($fieldName)
-	{
-		$defaultValues = array(
-			'username' => 'bigquery-service-account',
-			'password' => 'service-account-auth',
-			'db' => ''
-		);
-		$value = $defaultValues[$fieldName] ?? '';
-		return '<input type="hidden" name="auth[' . $fieldName . ']" value="' . htmlspecialchars($value) . '">' . "\n";
-	}
-
-	/**
-	 * OAuth2認証用のフィールドを描画
-	 */
-	private function renderOAuth2Field($name, $heading, $value)
-	{
-		switch ($name) {
-			case 'driver':
-				return $this->renderDriverField($heading);
-			case 'server':
-				return $this->renderOAuth2ProjectIdField();
-			case 'username':
-			case 'password':
-			case 'db':
-				return $this->renderHiddenField($name);
-			default:
-				return '';
-		}
-	}
-
-	/**
-	 * OAuth2認証用のProject IDフィールドを描画
-	 */
-	private function renderOAuth2ProjectIdField()
-	{
-		$default_value = htmlspecialchars($this->getProjectId());
-		return '<tr><th>Project ID</th><td><input name="auth[server]" value="' . $default_value . '" title="GCP Project ID" placeholder="your-project-id" autocapitalize="off" required></td></tr>' . "\n";
-	}
-
-	/**
-	 * Google OAuth2ログインボタンを描画
-	 */
-	private function renderOAuth2LoginButton()
-	{
-		$config = $this->getOAuth2Config();
-		$clientId = $config['client_id'];
-		$redirectUrl = $config['redirect_url'];
-
-		if (!$clientId || !$redirectUrl) {
-			return '<div class="oauth2-error">OAuth2 configuration incomplete. Please set GOOGLE_OAUTH2_CLIENT_ID and GOOGLE_OAUTH2_REDIRECT_URL.</div>';
-		}
-
-		// Google OAuth2 認証URL を構築
-		$scope = urlencode('https://www.googleapis.com/auth/bigquery https://www.googleapis.com/auth/cloud-platform');
-		$state = urlencode(base64_encode(json_encode(['redirect_to' => $_SERVER['REQUEST_URI']])));
-
-		$authUrl = 'https://accounts.google.com/o/oauth2/v2/auth?' . http_build_query([
-			'client_id' => $clientId,
-			'redirect_uri' => $redirectUrl,
-			'scope' => 'https://www.googleapis.com/auth/bigquery https://www.googleapis.com/auth/cloud-platform',
-			'response_type' => 'code',
-			'state' => $state,
-			'access_type' => 'offline',
-			'prompt' => 'consent'
-		]);
-
-		return '
-		<div class="oauth2-login-container">
-			<div class="oauth2-logo">
-				<svg width="40" height="40" viewBox="0 0 24 24" fill="none">
-					<path d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z" fill="#4285F4"/>
-					<path d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z" fill="#34A853"/>
-					<path d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z" fill="#FBBC05"/>
-					<path d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z" fill="#EA4335"/>
-				</svg>
-			</div>
-			<h2>Adminer</h2>
-			<a href="' . htmlspecialchars($authUrl) . '" class="oauth2-signin-button">Sign in with Google</a>
-		</div>';
-	}
-
-	function loginForm()
-	{
-		// OAuth2認証が有効な場合は専用のログイン画面を表示
-		if ($this->isOAuth2Enabled()) {
-			echo "<style>";
-			echo ".layout { display: none; }"; // 通常のログインフォームを非表示
-			echo ".oauth2-login-container {
-				display: flex;
-				flex-direction: column;
-				align-items: center;
-				justify-content: center;
-				min-height: 60vh;
-				text-align: center;
-				background: #f5f5f5;
-				margin: 20px;
-				border-radius: 8px;
-				padding: 40px;
-				box-shadow: 0 2px 10px rgba(0,0,0,0.1);
-			}";
-			echo ".oauth2-logo {
-				margin-bottom: 20px;
-			}";
-			echo ".oauth2-login-container h2 {
-				color: #5f6368;
-				font-size: 24px;
-				font-weight: 400;
-				margin: 10px 0 30px 0;
-			}";
-			echo ".oauth2-signin-button {
-				background: #1a73e8;
-				color: white;
-				border: none;
-				border-radius: 4px;
-				padding: 12px 24px;
-				font-size: 14px;
-				font-weight: 500;
-				text-decoration: none;
-				display: inline-block;
-				transition: background-color 0.2s;
-			}";
-			echo ".oauth2-signin-button:hover {
-				background: #1557b0;
-				color: white;
-				text-decoration: none;
-			}";
-			echo ".oauth2-error {
-				color: #d93025;
-				background: #fce8e6;
-				border: 1px solid #fce8e6;
-				border-radius: 4px;
-				padding: 12px 16px;
-				margin: 20px 0;
-			}";
-			echo "</style>";
-
-			// OAuth2ログインボタンを表示
-			echo $this->renderOAuth2LoginButton();
-		} else {
-			// 従来のCREDENTIAL認証用のスタイル
-			echo "<style>";
-			echo ".layout tr:has(input[type='hidden']) { display: none; }";
-			echo "</style>";
-		}
-	}
-
-	function operators()
-	{
-		return array(
-			"=", "!=", "<>", "<", "<=", ">", ">=",
-			"IN", "NOT IN", "IS NULL", "IS NOT NULL",
-			"LIKE", "NOT LIKE", "REGEXP", "NOT REGEXP"
-		);
-	}
-
-	protected $translations = array(
-		'en' => array('' => 'BigQuery authentication with service account credentials'),
-		'ja' => array('' => 'サービスアカウント認証情報によるBigQuery認証'),
-	);
-}
+add_driver("bigquery", "Google BigQuery");
 
 
 
-class AdminerBigQueryCSS extends Plugin
-{
-	private function isBigQueryDriver()
-	{
-		return (defined('DRIVER') && DRIVER === 'bigquery') || (defined('Adminer\\DRIVER') && constant('Adminer\\DRIVER') === 'bigquery');
-	}
 
-	function head($dark = null)
-	{
-		if ($this->isBigQueryDriver()) {
-
-			if (class_exists('Adminer\\Driver')) {
-				$driver = new \Adminer\Driver();
-				if (method_exists($driver, 'css')) {
-
-					echo $driver->css();
-				}
-			}
-		}
-	}
-}
 
 if (isset($_GET["bigquery"])) {
-	require_once __DIR__ . '/bigquery/BigQueryCacheManager.php';
-	require_once __DIR__ . '/bigquery/BigQueryConnectionPool.php';
 	define('Adminer\DRIVER', "bigquery");
 
-	require_once __DIR__ . '/bigquery/BigQueryConfig.php';
 
-	class BigQueryUtils
-	{
-
-		static function validateProjectId($projectId)
-		{
-			return preg_match('/^[a-z0-9][a-z0-9\\-]{4,28}[a-z0-9]$/i', $projectId) &&
-				strlen($projectId) <= 30;
-		}
-		static function escapeIdentifier($identifier)
-		{
-
-			if (preg_match('/^`[^`]*`$/', $identifier)) {
-				return $identifier;
-			}
-
-			$cleanIdentifier = trim($identifier, '`');
-			return "`" . str_replace("`", "``", $cleanIdentifier) . "`";
-		}
-		static function logQuerySafely($query, $context = "QUERY")
-		{
-			$sanitizers = array(
-				'/([\'"])[^\'\"]*\\1/' => '$1***REDACTED***$1',
-				'/\\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Z|a-z]{2,}\\b/' => '***EMAIL_REDACTED***'
-			);
-			$safeQuery = preg_replace(array_keys($sanitizers), array_values($sanitizers), $query);
-			if (strlen($safeQuery) > 200) {
-				$safeQuery = substr($safeQuery, 0, 200) . '... [TRUNCATED]';
-			}
-			error_log("BigQuery $context: $safeQuery");
-		}
-
-		static function convertValueForBigQuery($value, $fieldType)
-		{
-
-			if ($value === null) {
-				return 'NULL';
-			}
-
-			$cleanValue = trim(str_replace('`', '', $value));
-			$fieldType = strtolower($fieldType);
-
-			if (strpos($fieldType, 'timestamp') !== false) {
-				return "TIMESTAMP('" . str_replace("'", "''", $cleanValue) . "')";
-			} elseif (strpos($fieldType, 'datetime') !== false) {
-				return "DATETIME('" . str_replace("'", "''", $cleanValue) . "')";
-			} elseif (strpos($fieldType, 'date') !== false) {
-				return "DATE('" . str_replace("'", "''", $cleanValue) . "')";
-			} elseif (strpos($fieldType, 'time') !== false) {
-				return "TIME('" . str_replace("'", "''", $cleanValue) . "')";
-			} elseif (strpos($fieldType, 'int') !== false || strpos($fieldType, 'float') !== false || strpos($fieldType, 'numeric') !== false || strpos($fieldType, 'decimal') !== false) {
-
-				if (is_numeric($cleanValue)) {
-					return $cleanValue;
-				} else {
-					throw new InvalidArgumentException('Invalid numeric value: ' . $cleanValue);
-				}
-			} elseif (strpos($fieldType, 'bool') !== false) {
-				return (strtolower($cleanValue) === 'true' || $cleanValue === '1') ? 'TRUE' : 'FALSE';
-			} else {
-				return "'" . str_replace("'", "''", $cleanValue) . "'";
-			}
-		}
-
-		static function formatComplexValue($value, $field)
-		{
-			$fieldType = strtolower($field['type'] ?? 'text');
-			$typePatterns = array(
-				'json' => array('json', 'struct', 'record', 'array'),
-				'geography' => array('geography'),
-				'binary' => array('bytes', 'blob'),
-			);
-			foreach ($typePatterns as $handlerType => $patterns) {
-				if (self::matchesTypePattern($fieldType, $patterns)) {
-					return self::handleTypeConversion($value, $handlerType);
-				}
-			}
-			return $value;
-		}
-		private static function matchesTypePattern($fieldType, $patterns)
-		{
-			foreach ($patterns as $pattern) {
-				if (strpos($fieldType, $pattern) !== false) {
-					return true;
-				}
-			}
-			return false;
-		}
-		private static function handleTypeConversion($value, $handlerType)
-		{
-			switch ($handlerType) {
-				case 'json':
-					return is_string($value) && (substr($value, 0, 1) === '{' || substr($value, 0, 1) === '[')
-						? $value : json_encode($value);
-				case 'geography':
-				case 'binary':
-					return is_string($value) ? $value : (string) $value;
-				default:
-					return $value;
-			}
-		}
-		static function generateFieldConversion($field)
-		{
-
-			$fieldName = self::escapeIdentifier($field['field']);
-			$fieldType = strtolower($field['type'] ?? '');
-
-			// BigQuery固有データ型の変換マッピング
-			$conversions = array(
-				// 地理空間データの変換
-				'geography' => "ST_AsText($fieldName)",
-				'geom' => "ST_AsText($fieldName)",
-
-				// JSON・構造化データの変換
-				'json' => "TO_JSON_STRING($fieldName)",
-				'struct' => "TO_JSON_STRING($fieldName)",
-				'record' => "TO_JSON_STRING($fieldName)",
-				'array' => "TO_JSON_STRING($fieldName)",
-
-				// 日時データの変換
-				'timestamp' => "TIMESTAMP_TRUNC($fieldName, MICROSECOND)",
-				'datetime' => "DATETIME_TRUNC($fieldName, MICROSECOND)",
-				'time' => "TIME_TRUNC($fieldName, MICROSECOND)",
-
-				// バイナリデータの変換
-				'bytes' => "TO_BASE64($fieldName)",
-				'blob' => "TO_BASE64($fieldName)",
-
-				// 数値データの精度制御
-				'numeric' => "CAST($fieldName AS STRING)",
-				'bignumeric' => "CAST($fieldName AS STRING)",
-				'decimal' => "CAST($fieldName AS STRING)",
-
-				// 論理データの明示化
-				'boolean' => "IF($fieldName, 'true', 'false')",
-				'bool' => "IF($fieldName, 'true', 'false')"
-			);
-
-			// パターンマッチングで最適な変換を選択
-			foreach ($conversions as $typePattern => $conversion) {
-				if (strpos($fieldType, $typePattern) !== false) {
-					return $conversion;
-				}
-			}
-
-			// デフォルト: 変換不要
-			return null;
-		}
-
-		static function buildFullTableName($table, $database, $projectId)
-		{
-			return "`" . $projectId . "`.`" . $database . "`.`" . $table . "`";
-		}
-
-		/**
-		 * BigQueryジョブの完了状態を包括的に確認する共通関数
-		 *
-		 * @param object $job BigQueryジョブオブジェクト
-		 * @return bool ジョブが完了している場合はtrue
-		 */
-		static function isJobCompleted($job)
-		{
-			if (!$job) {
-				return false;
-			}
-
-			$jobInfo = $job->info();
-			$isJobComplete = false;
-
-			// 方法1: job->isComplete()メソッドによる確認
-			if ($job->isComplete()) {
-				$isJobComplete = true;
-			}
-
-			// 方法2: status.state フィールドによる確認
-			if (isset($jobInfo['status']['state']) && $jobInfo['status']['state'] === 'DONE') {
-				$isJobComplete = true;
-			}
-
-			// 方法3: statistics の存在による確認
-			if (isset($jobInfo['statistics'])) {
-				$isJobComplete = true;
-			}
-
-			return $isJobComplete;
-		}
-
-		/**
-		 * Process WHERE clause for BigQuery DML operations
-		 *
-		 * @param string $queryWhere The WHERE condition from Adminer
-		 * @return string Properly formatted WHERE clause with WHERE prefix
-		 * @throws InvalidArgumentException If WHERE condition is invalid
-		 */
-		static function processWhereClause($queryWhere)
-		{
-			if (empty($queryWhere) || trim($queryWhere) === '') {
-				return '';
-			}
-
-			$convertedWhere = convertAdminerWhereToBigQuery($queryWhere);
-
-			// Check if the converted WHERE already starts with WHERE keyword
-			if (preg_match('/^\s*WHERE\s/i', $convertedWhere)) {
-				return ' ' . $convertedWhere;
-			} else {
-				return ' WHERE ' . $convertedWhere;
-			}
-		}
-	}
+	// BigQueryUtils class is now in a separate file
+require_once __DIR__ . '/bigquery/BigQueryUtils.php';
 
 	if (!function_exists('Adminer\\idf_escape')) {
 		function idf_escape($idf)
